keep consistent manifest file name
removeBootstrappedManifestLocation is not called on exit, only when
disabling the addon. hence, restarting the browser would accumulate a
bunch of old manifests. onemen/117649 get around this by removing the
manifest on exit, but their code (following) is not working for me:
    Cc['@mozilla.org/uriloader/external-helper-app-service;1']
      .getService(Ci.nsPIExternalAppLauncher)
      .deleteTemporaryFileOnExit(tempDir);

partially reverts 7c2fc6e40e8e84d9cecde02e9630318ccca605ab.

additional note: our removeBootstrappedManifestLocation removes the
chrome.manifest (instead of truncating it with manifest.fileSize = 0).
this causes "Could not read chrome manifest 'file:///...'" to get
console.log'ged. this appears to be harmless and not worth fixing.